Refactor args
